+++ src/main/java/com/black_dog20/warpradial/Config.java
@@ -13,6 +13,11 @@ import java.nio.file.Path;
 public class Config {
 
     public static final String CATEGORY_GENERAL = "general";
+    public static final String CATEGORY_HOME = "home";
+    public static final String CATEGORY_SPAWN = "spawn";
+    public static final String CATEGORY_SERVER_WARPS = "serverWarp";
+    public static final String CATEGORY_PLAYER_WARPS = "playerWarp";
+    public static final String CATEGORY_ADMIN = "admin";
 
     private static final ForgeConfigSpec.Builder CLIENT_BUILDER = new ForgeConfigSpec.Builder();
     private static final ForgeConfigSpec.Builder SERVER_BUILDER = new ForgeConfigSpec.Builder();
@@ -37,40 +42,63 @@ public class Config {
     public static ForgeConfigSpec.IntValue COOLDOWN_PLAYER_WARP;
     public static ForgeConfigSpec.IntValue COOLDOWN_SERVER_WARP;
     public static ForgeConfigSpec.BooleanValue LOG_WARPS;
+    public static ForgeConfigSpec.BooleanValue ONLY_PERMISSION_PLAYERS_CAN_USE_MENU;
+    public static ForgeConfigSpec.BooleanValue INFORM_USER_OF_MISSING_PERMISSION;
 
     static {
 
-        CLIENT_BUILDER.comment("General settings").push(CATEGORY_GENERAL);
+        CLIENT_BUILDER.comment("General settings")
+                .push(CATEGORY_GENERAL);
         RADIAL_SCROLL_INVERTED = CLIENT_BUILDER.comment("Is scrolling inverted for the radial menu")
                 .define("invertedScroll", false);
         CLIENT_BUILDER.pop();
         CLIENT_CONFIG = CLIENT_BUILDER.build();
 
-        SERVER_BUILDER.comment("General settings").push(CATEGORY_GENERAL);
-        HOMES_ALLOWED = SERVER_BUILDER.comment("Can homes be created and warped to")
-                .define("homesAllowed", true);
-        WARP_TO_SPAWN_ALLOWED = SERVER_BUILDER.comment("Can spawn be warped to")
-                .define("spawnAllowed", true);
-        PLAYER_WARPS_ALLOWED = SERVER_BUILDER.comment("Can players create their own warps")
-                .define("playerWarpsAllowed", true);
-        SERVER_WARPS_ALLOWED = SERVER_BUILDER.comment("Can server warps be created and warped to")
-                .define("serverWarpsAllowed", true);
+        SERVER_BUILDER.comment("General settings")
+                .push(CATEGORY_GENERAL);
         COOLDOWN_MODE = SERVER_BUILDER.comment("Cooldown mode", "0 = no cooldown", "1 = global cooldown", "2 = per category cooldown")
                 .defineInRange("warpCooldownMode", PER_CATEGORY_COOLDOWN, 0, 2);
         COOLDOWN_GLOBAL = SERVER_BUILDER.comment("Global cooldown in ticks", "Is only used if cooldownMode = 1")
                 .defineInRange("globalWarpCooldownTicks", 2400, 0, 72000);
-        COOLDOWN_SPAWN_WARP = SERVER_BUILDER.comment("Spawn warp cooldown in ticks", "Is only used if warpCooldownMode = 2")
-                .defineInRange("spawnWarpCooldownTicks", 600, 0, 72000);
+
+        SERVER_BUILDER.push(CATEGORY_HOME);
+        HOMES_ALLOWED = SERVER_BUILDER.comment("Can homes be created and warped to")
+                .define("homesAllowed", true);
         COOLDOWN_HOME_WARP = SERVER_BUILDER.comment("Home warp cooldown in ticks", "Is only used if warpCooldownMode = 2")
                 .defineInRange("homeWarpCooldownTicks", 600, 0, 72000);
-        COOLDOWN_PLAYER_WARP = SERVER_BUILDER.comment("Player warps cooldown in ticks", "Is only used if warpCooldownMode = 2")
-                .defineInRange("playerWarpCooldownTicks", 2400, 0, 72000);
-        COOLDOWN_SERVER_WARP = SERVER_BUILDER.comment("Server warps cooldown in ticks", "Is only used if warpCooldownMode = 2")
-                .defineInRange("serverWarpCooldownTicks", 2400, 0, 72000);
-        LOG_WARPS = SERVER_BUILDER.comment("Are warp add or remove logged")
-                .define("logging", true);
         SERVER_BUILDER.pop();
 
+        SERVER_BUILDER.push(CATEGORY_SPAWN);
+        WARP_TO_SPAWN_ALLOWED = SERVER_BUILDER.comment("Can spawn be warped to")
+                .define("spawnAllowed", true);
+        COOLDOWN_SPAWN_WARP = SERVER_BUILDER.comment("Spawn warp cooldown in ticks", "Is only used if warpCooldownMode = 2")
+                .defineInRange("spawnWarpCooldownTicks", 600, 0, 72000);
+        SERVER_BUILDER.pop();
+
+        SERVER_BUILDER.push(CATEGORY_PLAYER_WARPS);
+        PLAYER_WARPS_ALLOWED = SERVER_BUILDER.comment("Can players create their own warps")
+                .define("playerWarpsAllowed", true);
+        COOLDOWN_PLAYER_WARP = SERVER_BUILDER.comment("Player warps cooldown in ticks", "Is only used if warpCooldownMode = 2")
+                .defineInRange("playerWarpCooldownTicks", 2400, 0, 72000);
+        SERVER_BUILDER.pop();
+
+        SERVER_BUILDER.push(CATEGORY_SERVER_WARPS);
+        SERVER_WARPS_ALLOWED = SERVER_BUILDER.comment("Can server warps be created and warped to")
+                .define("serverWarpsAllowed", true);
+        COOLDOWN_SERVER_WARP = SERVER_BUILDER.comment("Server warps cooldown in ticks", "Is only used if warpCooldownMode = 2")
+                .defineInRange("serverWarpCooldownTicks", 2400, 0, 72000);
+        SERVER_BUILDER.pop();
+
+        SERVER_BUILDER.push(CATEGORY_ADMIN);
+        LOG_WARPS = SERVER_BUILDER.comment("Are warp add or remove logged")
+                .define("logging", true);
+        ONLY_PERMISSION_PLAYERS_CAN_USE_MENU = SERVER_BUILDER.comment("Only allow ops and players with specific permission to use the radial menu")
+                .define("onlyOpsMenu", false);
+        INFORM_USER_OF_MISSING_PERMISSION = SERVER_BUILDER.comment("Inform the player that they do not have the required permission to use the menu")
+                .define("tellUsersOpsMenu", true);
+        SERVER_BUILDER.pop();
+
+        SERVER_BUILDER.pop();
         SERVER_CONFIG = SERVER_BUILDER.build();
     }
 
@@ -79,6 +107,7 @@ public class Config {
         final CommentedFileConfig configData = CommentedFileConfig.builder(path)
                 .sync()
                 .autosave()
+                .preserveInsertionOrder()
                 .writingMode(WritingMode.REPLACE)
                 .build();