+++ CHANGELOG.md
@@ -1,7 +1,15 @@
 # Warp Radial Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
-## [1.15.2, 1.0.0] - ????
+## [1.15.2, 1.0.1] - 2020-09-26
+- Fixes issue with can delete requiring permission to create instead of delete.
+- Changed the config layout.
+- Added new config setting to allow only ops and specific players to use the radial menu.
+- Added new permission: canUseMenu
+- Added command to give players permissions to use the radial menu.
+- Fixed permission saving issue.
+
+## [1.15.2, 1.0.0] - 2020-09-05
 - The whole mod!
 
 # Example